+++ src/polynomials.rs
@@ -53,7 +53,7 @@ pub mod evaluation {
     /// This function evaluates a polynomial and its derivatives storing the results in the array res of size lenres. The output array contains
     /// the values of d^k P/d x^k for the specified value of x starting with k = 0.
     pub fn poly_eval_derivs(c: &[f64], x: f64, res: &mut [f64]) -> enums::value::Value {
-        unsafe { ffi::gsl_poly_eval_derivs(c.as_ptr(), c.len() as u64, x, res.as_mut_ptr(), res.len() as u64) }
+        unsafe { ffi::gsl_poly_eval_derivs(c.as_ptr(), c.len() as usize, x, res.as_mut_ptr(), res.len() as usize) }
     }
 }
 
@@ -81,19 +81,19 @@ pub mod divided_difference_representation {
     /// xa and ya of length size. On output the divided-differences of (xa,ya) are stored in the array dd, also of length size. Using the
     /// notation above, dd[k] = [x_0,x_1,...,x_k].
     pub fn poly_dd_init(dd: &mut [f64], xa: &[f64], ya: &[f64]) -> enums::value::Value {
-        unsafe { ffi::gsl_poly_dd_init(dd.as_mut_ptr(), xa.as_ptr(), ya.as_ptr(), dd.len() as u64) }
+        unsafe { ffi::gsl_poly_dd_init(dd.as_mut_ptr(), xa.as_ptr(), ya.as_ptr(), dd.len() as usize) }
     }
 
     /// This function evaluates the polynomial stored in divided-difference form in the arrays dd and xa of length size at the point x.
     pub fn poly_dd_eval(dd: &[f64], xa: &[f64], x: f64) -> f64 {
-        unsafe { ffi::gsl_poly_dd_eval(dd.as_ptr(), xa.as_ptr(), dd.len() as u64, x) }
+        unsafe { ffi::gsl_poly_dd_eval(dd.as_ptr(), xa.as_ptr(), dd.len() as usize, x) }
     }
 
     /// This function converts the divided-difference representation of a polynomial to a Taylor expansion. The divided-difference representation
     /// is supplied in the arrays dd and xa of length size. On output the Taylor coefficients of the polynomial expanded about the point xp are
     /// stored in the array c also of length size. A workspace of length size must be provided in the array w.
     pub fn poly_dd_taylor(c: &mut [f64], xp: f64, dd: &[f64], xa: &[f64], w: &mut [f64]) -> enums::value::Value {
-        unsafe { ffi::gsl_poly_dd_taylor(c.as_mut_ptr(), xp, dd.as_ptr(), xa.as_ptr(), dd.len() as u64, w.as_mut_ptr()) }
+        unsafe { ffi::gsl_poly_dd_taylor(c.as_mut_ptr(), xp, dd.as_ptr(), xa.as_ptr(), dd.len() as usize, w.as_mut_ptr()) }
     }
 
     /// This function computes a divided-difference representation of the interpolating Hermite polynomial for the points (x, y) stored in the
@@ -103,7 +103,7 @@ pub mod divided_difference_representation {
     /// of the Hermite representation are stored in the array dd, also of length 2*size. Using the notation above, dd[k] = [z_0,z_1,...,z_k].
     /// The resulting Hermite polynomial can be evaluated by calling gsl_poly_dd_eval and using za for the input argument xa.
     pub fn poly_dd_hermite_init(dd: &mut [f64], za: &mut [f64], xa: &[f64], ya: &[f64], dya: &[f64]) -> enums::value::Value {
-        unsafe { ffi::gsl_poly_dd_hermite_init(dd.as_mut_ptr(), za.as_mut_ptr(), xa.as_ptr(), ya.as_ptr(), dya.as_ptr(), dd.len() as u64) }
+        unsafe { ffi::gsl_poly_dd_hermite_init(dd.as_mut_ptr(), za.as_mut_ptr(), xa.as_ptr(), ya.as_ptr(), dya.as_ptr(), dd.len() as usize) }
     }
 }
